Initial creation of admin pages; creation of email contact form
Began task of creating set of "admin" pages for db updates. Needs
"users" table in db. Also created contact form in person.php template,
which automatically sets the subject to the name and ID of the person
shown on the page. Still has kinks for working out, which is why it is
being committed to "development" instead of  "master".