Add Nav Buttons to Edit Page
I added smart navigation buttons to the edit_site.php page. If you are
on site ID 1, you will not see the "previous" button. If you are on the
last page (as determined by a COUNT(*) query), you will not see the
"next" button. All other pages display both a "previous" and a "next
button", and there is a set at the top and the bottom of the page. The
preexisting links back to the sites list and to the next page needing
verification have been incorporated into the same button set. The link
to the next unverified page seems to be broken - it always brings you to
site ID 1, despite the fact that I updated the site locally and a date
stamp shows in the DB. Maybe the query needs to only SELECT rows where
site_verified == NULL.
BUG: For some reason, three double quotes are being inserted into the
HTML after the top buttons and before the table is opened. They appear
as one double-quote on the page, directly to the right of the button
row. In the source, there are three. I can't find the cause of it in the
PHP.