Initial Implementation of geocode function
geocode() function currently returns 0 and the rest of the function is
only pseudocode. Needs JS and some data conversion.
edit_page has commented-out section calling the geocode() function if
lat or lng is null, storing the return of geocode() into the appropriate
variables. Not sure if there's a better place to put this function,
given that it relies on variables created while creating the lat and
long table rows, so we can't call it before the table is created.
TODO: need to obtain a GA Google Maps API key. Should ideally be tied to
a Google account owned by the Kingdom, vs a personal Google account.
That way the key stays with the Kingdom.