Geocoding, Session Expiry, and Aesthetics
Geocoding function now works, and inserts coordinates into coordinate
fields any time lat or lng is not set and street is set. Displays a
notice requesting a page submit to write the data to the db.
Session expiry now appears to be functioning properly. There is a set of
echo statements displaying the contents of the session expiry variables
so you can monitor them to see if they are updating properly, and for
testing to see if the 3 hour expiration works.
edit_site.php has had the main column reduced to 6 wide for medium and
larger viewports, and at 12 wide for small and x-small viewports. Can
test on mobile devices and adjust these as needed.