Session Validate Constants; Search styling; Sample db
Relocated Session Validate time constants to the constants file. Made
some small changes to search.php (links to sections now all in one row,
groups section subheader now refers to the number of groups found,
instead of the number of awards found. Sanitized a copy of the current
db in dev.records and placed in / as a sample of the db structure for
other Kingdoms to use. No personal data contained therein.