Readability Improvements (Event Page)
Event page includes clicky links to event site and hosting group. Events
with an "unknown" location will link to site "-1" which will then
display a custom error message on site.php. Error message encourages the
visitor to report the error to the webminister, which we can then pass
along to the Heralds who are responsible for updating the information.