First API script page: Person Awards
This script takes one variable as input via $_GET, and returns as a
result the associated person's name and awards. The result is echoed to
the page as JSON encoded text, and can be retrieved via cURL through php
on external sites. The script does do error checking to ensure that the
passed in variable is numeric. It will be up to the author of the
requesting page to write in any error checking in the event that the
supplied numeric variable returns a failure.