Main1 Template plus footer proposal
This commit contains both tweaks to the appearance and content of Main1,
and also changes to the global footer file.
Column headers in Awards/Marshals/Campgrounds/Website descriptions are
now images. Found a tent icon that goes reasonably well with the rest.
Also bolded a keyword in each column to act as an inline column category
identifier for those who don't read devices. Image alt texts include
blazon for the three images that are devices. For the tent icon, it
says, "(Unofficial) Red and white striped pavilion".
Feature boxes adjusted to be roughly same height, and search box has
been broken out of these and combined into one box just below that row.
Since having search forms in both boxes are actually the exact same
search form, with the exact same results given identical input, it seems
it would be counter intuitive to expect users to assume that the two
search boxes do the exact same thing. So, I moved it to a location
centered directly below the feature boxes to tie everything together.
Feature request: include campground name matches on the search results
list?
Global footer now has a black bar at the very bottom with a left-aligned
"Copyright xxxx-xxxx" paragraph which starts with 2015 and ends with the
current date's year (via php).
Directly above the black bar is a red bar (same color as header) with
the disclaimers in 8 columns on medium and up devices, and 12 columns in
xs and s devices, followed by social media link icons for Kingdom SM
properties and then the Kingdom device (mostly to fill empty space)