Add printable authorization card
This page produces a printer-friendly authorization card. $_GET data of
?id=xxx must be provided, where xxx is the id number of the individual
whose card you want to produce. If errors involving authorizations not
displaying when they should, most likely place to check first is
spelling of auth or warrant name in code vs spelling in DB, as this file
uses strcmp operations to determine whether or not to store a value in
$authorizations or $warrants.