Displaying images retrieved from database
edit_person_sub_armorial.php now properly displays images from query
results. Also created armorial.php in public and templates. Was created
as a way to easily troubleshoot the issue with displaying images.
Included as a proposed "gallery" page. Ultimately I envision a grid of
device images, with the blazon as the hovertext, and each one links to
its owner. Could be set to only display personal devices, and not
badges. Flexible.