Page layout redesign (draft)
A proposed redesign, featuring the following elements

* device image and badge images display in a row below name of person
and paper copy disclaimer. The columns are set to allow the images to
stay near the center of the page via column offsets. The more images,
the smaller the offset.
* Awards and combat info are displayed side by side in medium or larger
viewports
* a "Skip to Awards" link is available at the bottom of the name header
to allow those on mobile devices to skip the images, which will stack
vertically in sm and xs viewports.